Spherical harmonics seem to need to have the normal inverted

+++ src/rendering/Deferred.cpp
@@ -20,7 +20,7 @@ const Extent2D DeferredRenderer::hbao_extent(512, 512);
 const Extent2D DeferredRenderer::ssr_extent(512, 512);
 
 DeferredPass::DeferredPass(bool is_indirect_pass)
-    : FullScreenPass(InternalFormat::RGBA16F),
+    : FullScreenPass(InternalFormat::RGBA32F),
       m_is_indirect_pass(is_indirect_pass)
 {
 }
